add format indicator to get LOG_X warnings
The LOG_X macros emit a lot of warnings when compiling with the T
because they call the function logRecord (or logRecord_mt)
which has not been "marked" as calling printf, so gcc won't
emit typical printf warnings.

With the T, they directly translate to sprintf, so those warnings
pop up, much more verbose due to several macro expansions.

Let's make them pop up all the time so it's easier to fix them.