bugfix: dlsim was not working anymore
When you ran it with the scope as:

    ./dlsim -X

you saw no activity in the "PDSCH I/Q of MF Output" window,
so matter what SNR you put (even for very high SNR, say: ./dlsim -X -s40)

this commit seems to fix it.

I am not sure it is the end of story. The 2 threads PDCCH/PDSCH structures
may still not be used correctly everywhere.