hotfix: do not set s-Measure in measurement report configuration
The value was set to 20 which means -120 dB. Setting this value means
that the UE reports an event only if the RSRP of the serving cell
is lower than -120 dB.

In the lab, we get values of around -70, and to go as low as -120
means a very degraded environment. This is not practical.

Let's disactivate the setting of this parameter for the moment so that
the UE reports an event without taking into consideration the RSRP
of the serving cell.

With this hotfix, the X2 handover works much better.