lte-softmodem: fix if4p5 (maybe incorrect)
For whatever reason, ru->frame_parms was NULL and the RRU crashed.
Let's allocated memory.

Also, don't do fp = ru->frame_parms in ru_thread, it seems that it
takes the value before allocation (so: NULL) and the code crashes
later on.

Maybe it's not the correct solution. To be refined at some point.