HO is done on reconfigurationComplete
We should remove the HO_context thing as soon as we receive
reconfiguration complete: it matches N2 (then we only update the source
gNB, for the target gNB, HO is done). Further, we should within the CU
keep track with a flag if the AMF requested the release, and only free
the UE context at RRC if this is the case.